backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-community
oci
oracledb
pydantic
pydantic-settings
python-dotenv
requests
